Documented Writing Kernels section
