--- numba_dpex/examples/sum.py
@@ -1,54 +1,0 @@
-#! /usr/bin/env python
-
-# SPDX-FileCopyrightText: 2020 - 2022 Intel Corporation
-#
-# SPDX-License-Identifier: Apache-2.0
-
-import dpctl
-import numpy as np
-import numpy.testing as testing
-
-import numba_dpex as dpex
-
-
-@dpex.kernel
-def data_parallel_sum(a, b, c):
-    """
-    Vector addition using the ``kernel`` decorator.
-    """
-    i = dpex.get_global_id(0)
-    c[i] = a[i] + b[i]
-
-
-def driver(a, b, c, global_size):
-    print("A : ", a)
-    print("B : ", b)
-    data_parallel_sum[global_size, dpex.DEFAULT_LOCAL_SIZE](a, b, c)
-    print("A + B = ")
-    print("C ", c)
-    testing.assert_equal(c, a + b)
-
-
-def main():
-    global_size = 10
-    N = global_size
-    print("N", N)
-
-    a = np.array(np.random.random(N), dtype=np.float32)
-    b = np.array(np.random.random(N), dtype=np.float32)
-    c = np.ones_like(a)
-
-    # Use the environment variable SYCL_DEVICE_FILTER to change the default device.
-    # See https://github.com/intel/llvm/blob/sycl/sycl/doc/EnvironmentVariables.md#sycl_device_filter.
-    device = dpctl.select_default_device()
-    print("Using device ...")
-    device.print_device_info()
-
-    with dpctl.device_context(device):
-        driver(a, b, c, global_size)
-
-    print("Done...")
-
-
-if __name__ == "__main__":
-    main()
